db create random data
file searcher

+++ java/java.helpers/src/main/java/com/mtsmda/generator/GenerateRandom.java
@@ -190,7 +190,7 @@ public class GenerateRandom {
     }
 
     public static int generateRandomNumberInRange(int min, int max) {
-        return generateRandomNumberInRange(min, max);
+        return (int) generateRandomNumberInRange((long) min, (long)max);
     }
 
     public static long generateRandomNumberInRange(long min, long max) {
